requirements.txt
Flask
Flask-Cors
psycopg2-binary
python-dotenv
argon2-cffi